:recycle: Optimise check duplication for public keys functions
