+++ src/transactions/3_castVotes.js
@@ -19,7 +19,6 @@
 import cryptoModule from '../crypto';
 import { VOTE_FEE } from '../constants';
 import {
-	hasNoDuplication,
 	prepareTransaction,
 	getTimeWithOffset,
 	prependMinusToPublicKeys,
@@ -46,7 +45,6 @@ const castVotes = ({
 	secondPassphrase,
 	timeOffset,
 }) => {
-	hasNoDuplication([...(votes || []), ...(unvotes || [])]);
 	const keys = cryptoModule.getKeys(passphrase);
 
 	validatePublicKeys([...votes, ...unvotes]);
