+++ src/transactions/4_registerMultisignatureAccount.js
@@ -15,7 +15,6 @@
 import cryptoModule from '../crypto';
 import { MULTISIGNATURE_FEE } from '../constants';
 import {
-	hasNoDuplication,
 	prepareTransaction,
 	getTimeWithOffset,
 	prependPlusToPublicKeys,
@@ -42,7 +41,6 @@ const registerMultisignatureAccount = ({
 	minimum,
 	timeOffset,
 }) => {
-	hasNoDuplication(keysgroup);
 	const keys = cryptoModule.getKeys(passphrase);
 
 	validateKeysgroup(keysgroup);
