+++ test/transactions/utils/validation.js
@@ -14,6 +14,7 @@
  */
 import cryptoModule from '../../../src/crypto';
 import {
+	checkPublicKeysForDuplicates,
 	validatePublicKey,
 	validatePublicKeys,
 	validateKeysgroup,
@@ -153,6 +154,34 @@ describe('public key validation', () => {
 					.bind(null, keysgroup)
 					.should.throw(
 						'Expected between 1 and 16 public keys in the keysgroup.',
+					);
+			});
+		});
+	});
+
+	describe('#checkPublicKeysForDuplicates', () => {
+		describe('Given an array of public keys without duplication', () => {
+			const publicKeys = [
+				'215b667a32a5cd51a94c9c2046c11fffb08c65748febec099451e3b164452bca',
+				'922fbfdd596fa78269bbcadc67ec2a1cc15fc929a19c462169568d7a3df1a1aa',
+				'5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09',
+			];
+			it('should return true', () => {
+				return checkPublicKeysForDuplicates(publicKeys).should.be.true();
+			});
+		});
+
+		describe('Given an array of public keys with duplication', () => {
+			const publicKeys = [
+				'215b667a32a5cd51a94c9c2046c11fffb08c65748febec099451e3b164452bca',
+				'922fbfdd596fa78269bbcadc67ec2a1cc15fc929a19c462169568d7a3df1a1aa',
+				'922fbfdd596fa78269bbcadc67ec2a1cc15fc929a19c462169568d7a3df1a1aa',
+			];
+			it('should throw', () => {
+				return checkPublicKeysForDuplicates
+					.bind(null, publicKeys)
+					.should.throw(
+						'Duplicated public key: 922fbfdd596fa78269bbcadc67ec2a1cc15fc929a19c462169568d7a3df1a1aa.',
 					);
 			});
 		});
